+++ lib/src/extensions/objects.dart
@@ -9,7 +9,7 @@ extension ObjectExtensions on Object? {
   ///  * Object is true only if
   ///    1. Object is bool and true.
   ///    2. Object is num and is greater than zero.
-  ///    3. Object is string and is equal to 'yes', 'true', or '1'.
+  ///    3. Object is string and is equal to 'yes', 'true', '1', or 'ok'.
   /// * any other conditions including null will return false.
   bool get asBool {
     if (this is bool?) return (this as bool?) ?? false;