+++ lib/src/extensions/string_ext.dart
@@ -371,7 +371,11 @@ extension NullSafeStringExtensions on String? {
     try {
       final s =
           this?.removeWhiteSpaces.removeEmptyLines.toLowerCase() ?? 'false';
-      return s == 'true' || s == 'yes' || s == '1' || s.tryToNum.asBool;
+      return s == 'true' ||
+          s == 'yes' ||
+          s == '1' ||
+          s == 'ok' ||
+          s.tryToNum.asBool;
     } catch (_) {
       return false;
     }